refactor: streamline menu ingredient amount edit flow
Refactor ingredient edit screen to focus on recipe amount updates only.

Update menu recipe amount contracts to Double for number-based PATCH requests.

Remove unused add bottom sheet and add IngredientEditViewModel unit tests.

+++ app/src/main/java/com/team/chord/navigation/MenuAddFlowViewModel.kt
@@ -118,7 +118,7 @@ class MenuAddFlowViewModel @Inject constructor(
                         menuId = 0L,
                         ingredientId = ingredient.id,
                         ingredientName = ingredient.name,
-                        amount = ingredient.amount,
+                        amount = ingredient.amount.toDouble(),
                         unitCode = ingredient.unit.name,
                         price = ingredient.price,
                     )
