+++ core/core-network/src/main/kotlin/com/team/chord/core/network/mapper/MenuMapper.kt
@@ -61,7 +61,7 @@ fun RecipeDto.toDomain(): MenuRecipe =
         menuId = menuId,
         ingredientId = ingredientId,
         ingredientName = ingredientName,
-        amount = amount.toInt(),
+        amount = amount,
         unitCode = unitCode,
         price = price.toInt(),
     )
